test(ci): separate classification and ownership audits
